+++ data/scenarios/Challenges/_pack-tetrominoes/solution.sw
@@ -1,4 +1,3 @@
-
 move;
 place "j-tile";
 move;
